+++ Example/nRFMeshProvision/View Controllers/Network/Configuration/NodeAppKeysViewController.swift
@@ -46,7 +46,9 @@ class NodeAppKeysViewController: ProgressViewController, Editable {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        tableView.setEmptyView(title: "No keys", message: "Click + to add a new key.", messageImage: #imageLiteral(resourceName: "baseline-key"))
+        tableView.setEmptyView(title: "No keys",
+                               message: "Click + to add a new key.",
+                               messageImage: #imageLiteral(resourceName: "baseline-key"))
     }
     
     override func viewWillAppear(_ animated: Bool) {