+++ Example/nRFMeshProvision/View Controllers/Network/Provisioning/ScannerTableViewController.swift
@@ -61,7 +61,9 @@ class ScannerTableViewController: UITableViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        tableView.setEmptyView(title: "Can't see your device?", message: "1. Make sure the device is turned on\nand connected to a power source.\n\n2. Make sure the relevant firmware\nand SoftDevices are flashed.", messageImage: #imageLiteral(resourceName: "baseline-bluetooth"))
+        tableView.setEmptyView(title: "Can't see your device?",
+                               message: "1. Make sure the device is turned on\nand connected to a power source.\n\n2. Make sure the relevant firmware\nand SoftDevices are flashed.",
+                               messageImage: #imageLiteral(resourceName: "baseline-bluetooth"))
         centralManager = CBCentralManager()
         
         tableView.showEmptyView()